feat: create a unified error & error response structure
An interface 'IBaseError' was created - this is the expected structure
for all errors thrown in this application.
A reference implementation (`BaseError`) was also created with a
convienient constructor that allows each `IBaseError` property to be
defined on construction.
Multiple verification an casting functions were created to verify that
errors conform to the created specification and reconstruct them if they
violate it.

The error handler of the whole application was rewritten to use this
error handling framework instead of the default.