feat: rust-inspired error wrapping method for Promise
This commit adds a `.addErrorContext()` method on the `Promise` class,
inspired by rust's `anyhow` library and its `.context`/`.with_context`
methods.
Due to the way JS handles errors, this method can only be viably
implemented on Promises, but that still allows it to be used in many
ways, such as to modify error messages from `.*OrFail()` database
operations.
This method takes in either a `ContextSpec` value, or a function that
returns `ContextSpec` and wraps any thrown errors according to that
value.
Functions passed to `.addErrorContext()` are lazily evaluated only if an
error is thrown, and as such they should be used if the spec value is
calculated based on variable arguments, such when the context message
includes the query values.

This feature is then immediately used on most controllers, to set better
error messages when the queried ID does not exist.