refactor: unify error handling (#132)
* feat: create a unified error & error response structure

An interface 'IBaseError' was created - this is the expected structure
for all errors thrown in this application.
A reference implementation (`BaseError`) was also created with a
convienient constructor that allows each `IBaseError` property to be
defined on construction.
Multiple verification an casting functions were created to verify that
errors conform to the created specification and reconstruct them if they
violate it.

The error handler of the whole application was rewritten to use this
error handling framework instead of the default.

* refactor: use the new error handling framework where applicable

* feat: rust-inspired error wrapping method for Promise

This commit adds a `.addErrorContext()` method on the `Promise` class,
inspired by rust's `anyhow` library and its `.context`/`.with_context`
methods.
Due to the way JS handles errors, this method can only be viably
implemented on Promises, but that still allows it to be used in many
ways, such as to modify error messages from `.*OrFail()` database
operations.
This method takes in either a `ContextSpec` value, or a function that
returns `ContextSpec` and wraps any thrown errors according to that
value.
Functions passed to `.addErrorContext()` are lazily evaluated only if an
error is thrown, and as such they should be used if the spec value is
calculated based on variable arguments, such when the context message
includes the query values.

This feature is then immediately used on most controllers, to set better
error messages when the queried ID does not exist.

* fix(controllers/files): accept keys with file extensions

whoopsies

* fix: ok, maybe eslint was right... that one time (out of 4)