feat(scrapers): conditional execution
this commit adds a shouldRun() method to scraper modules, which allows
them to decide whether they should be run.
this will allow for the safe usage of --run-all on an existing database,
as the modules that don't need to be run again won't be run again.

this check can be skipped using the --force flag