feat(scrapers): conditional exectuion & module selection by cmdline (#192)
* refactor(exceptions): split out the error logging code to a function

* refactor(scrapers): use the error handling framework for error handling

* chore: bump @poppinss/cliui and remove @ts-expect-error

* feat(exceptions): shorten stack traces by trimming out CWD

* feat(scrapers): conditional execution

this commit adds a shouldRun() method to scraper modules, which allows
them to decide whether they should be run.
this will allow for the safe usage of --run-all on an existing database,
as the modules that don't need to be run again won't be run again.

this check can be skipped using the --force flag

* feat(scrapers): allow selection of modules via commandline args

you can now put module names or file names after the db:scrape command.
only those modules will be run, with no selection prompt

* feat(scrapers): implement shouldRun() on all existing scrapers